Разделил файл utils.js на несколько
